move getFinalScreenshot to PSI
