+++ lighthouse-core/test/report/html/renderer/psi-test.js
@@ -14,15 +14,18 @@ const URL = require('../../../../lib/url-shim');
 const PSI = require('../../../../report/html/renderer/psi.js');
 const Util = require('../../../../report/html/renderer/util.js');
 const DOM = require('../../../../report/html/renderer/dom.js');
-const CategoryRenderer =
-    require('../../../../report/html/renderer/category-renderer');
+const CategoryRenderer = require('../../../../report/html/renderer/category-renderer');
 const DetailsRenderer = require('../../../../report/html/renderer/details-renderer');
 const CriticalRequestChainRenderer =
     require('../../../../report/html/renderer/crc-details-renderer');
 
 const sampleResultsStr = fs.readFileSync(__dirname + '/../../../results/sample_v2.json', 'utf-8');
-const TEMPLATE_FILE = fs.readFileSync(__dirname +
-  '/../../../../report/html/templates.html', 'utf8');
+const sampleResults = JSON.parse(sampleResultsStr)
+;
+const TEMPLATE_FILE = fs.readFileSync(
+  __dirname + '/../../../../report/html/templates.html',
+  'utf8'
+);
 
 /* eslint-env jest */
 
@@ -54,47 +57,69 @@ describe('DOM', () => {
     global.CriticalRequestChainRenderer = undefined;
   });
 
-  describe('psi prepareLabData helper', () => {
-    it('reports expected data', () => {
-      const result = PSI.prepareLabData(sampleResultsStr, document);
-      assert.ok(result.scoreGaugeEl instanceof document.defaultView.Element);
-      assert.ok(result.perfCategoryEl instanceof document.defaultView.Element);
-      assert.equal(typeof result.finalScreenshotDataUri, 'string');
+  describe('psi prepareLabData helpers', () => {
+    describe('prepareLabData', () => {
+      it('reports expected data', () => {
+        const result = PSI.prepareLabData(sampleResultsStr, document);
+        assert.ok(result.scoreGaugeEl instanceof document.defaultView.Element);
+        assert.ok(result.perfCategoryEl instanceof document.defaultView.Element);
+        assert.equal(typeof result.finalScreenshotDataUri, 'string');
 
-      assert.ok(result.finalScreenshotDataUri.startsWith('data:image/jpeg;base64,'));
-      assert.ok(result.scoreGaugeEl.outerHTML.includes('<style>'), 'score gauge comes with CSS');
-      assert.ok(result.scoreGaugeEl.outerHTML.includes('<svg'), 'score gauge comes with SVG');
-      assert.ok(result.perfCategoryEl.outerHTML.length > 50000, 'perfCategory HTML is populated');
+        assert.ok(result.finalScreenshotDataUri.startsWith('data:image/jpeg;base64,'));
+        assert.ok(result.scoreGaugeEl.outerHTML.includes('<style>'), 'score gauge comes with CSS');
+        assert.ok(result.scoreGaugeEl.outerHTML.includes('<svg'), 'score gauge comes with SVG');
+        assert.ok(result.perfCategoryEl.outerHTML.length > 50000, 'perfCategory HTML is populated');
+      });
+
+      it('throws if there is no perf category', () => {
+        const lhrWithoutPerf = JSON.parse(sampleResultsStr);
+        delete lhrWithoutPerf.categories.performance;
+        const lhrWithoutPerfStr = JSON.stringify(lhrWithoutPerf);
+
+        assert.throws(() => {
+          PSI.prepareLabData(lhrWithoutPerfStr, document);
+        }, /no performance category/i);
+      });
+
+      it('throws if there is no perf category', () => {
+        const lhrWithoutPerf = JSON.parse(sampleResultsStr);
+        delete lhrWithoutPerf.categories.performance;
+        const lhrWithoutPerfStr = JSON.stringify(lhrWithoutPerf);
+
+        assert.throws(() => {
+          PSI.prepareLabData(lhrWithoutPerfStr, document);
+        }, /no performance category/i);
+      });
+
+      it('throws if there is no category groups', () => {
+        const lhrWithoutGroups = JSON.parse(sampleResultsStr);
+        delete lhrWithoutGroups.categoryGroups;
+        const lhrWithoutGroupsStr = JSON.stringify(lhrWithoutGroups);
+
+        assert.throws(() => {
+          PSI.prepareLabData(lhrWithoutGroupsStr, document);
+        }, /no category groups/i);
+      });
+    });
+  });
+
+  describe('getFinalScreenshot', () => {
+    it('gets a datauri as a string', () => {
+      const cloneResults = Util.prepareReportResult(sampleResults);
+      const perfCategory = cloneResults.reportCategories.find(cat => cat.id === 'performance');
+      const datauri = PSI.getFinalScreenshot(perfCategory);
+      assert.equal(typeof datauri, 'string');
+      assert.ok(datauri.startsWith('data:image/jpeg;base64,'));
     });
 
-    it('throws if there is no perf category', () => {
-      const lhrWithoutPerf = JSON.parse(sampleResultsStr);
-      delete lhrWithoutPerf.categories.performance;
-      const lhrWithoutPerfStr = JSON.stringify(lhrWithoutPerf);
+    it('returns null if there is no final-screenshot audit', () => {
+      const clonedResults = JSON.parse(JSON.stringify(sampleResults));
+      delete clonedResults.audits['final-screenshot'];
+      const lhrNoFinalSS = Util.prepareReportResult(clonedResults);
+      const perfCategory = lhrNoFinalSS.reportCategories.find(cat => cat.id === 'performance');
 
-      assert.throws(() => {
-        PSI.prepareLabData(lhrWithoutPerfStr, document);
-      }, /no performance category/i);
-    });
-
-    it('throws if there is no perf category', () => {
-      const lhrWithoutPerf = JSON.parse(sampleResultsStr);
-      delete lhrWithoutPerf.categories.performance;
-      const lhrWithoutPerfStr = JSON.stringify(lhrWithoutPerf);
-
-      assert.throws(() => {
-        PSI.prepareLabData(lhrWithoutPerfStr, document);
-      }, /no performance category/i);
-    });
-
-    it('throws if there is no category groups', () => {
-      const lhrWithoutGroups = JSON.parse(sampleResultsStr);
-      delete lhrWithoutGroups.categoryGroups;
-      const lhrWithoutGroupsStr = JSON.stringify(lhrWithoutGroups);
-
-      assert.throws(() => {
-        PSI.prepareLabData(lhrWithoutGroupsStr, document);
-      }, /no category groups/i);
+      const datauri = PSI.getFinalScreenshot(perfCategory);
+      assert.equal(datauri, null);
     });
   });
 });
